requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.30.6

# OpenAI for LLM and embeddings
openai==1.55.3

# Graph processing
networkx==3.2.1

# Data validation and utilities
pydantic==2.11.9
python-dotenv==1.1.1